requirements.txt
# Web Framework
fastapi[standard]>=0.115.0
uvicorn>=0.30.0
pydantic-settings>=2.4.0

# AI & LLM
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
huggingface-hub>=0.20.0
langchain-openai>=0.1.0
groq>=0.5.0

# Web Search
tavily-python>=0.3.0

# Document Parsing (CV)
pdfplumber>=0.11.0

# Utilities
python-multipart>=0.0.9
python-dotenv>=1.0.1
requests>=2.31.0